Produce a single static library with all VQE code
Previously VQE builds several static libraries some of which rely on
symbols from others.  This means that these libraries must be linked
in a specific order.  This can be inconvenient as both pkg-config and
libtool will re-order link arguments, causing linker failures.

With this commit a static library 'libvqecuber' is produced which
wont cause these issues.