Fix build on x86_64
This just sets up the defines identically such that vqe-c will treat
x86_64 identically to i386 so the build stops failing on x86_64.  The
real fix would be to make it do something sensible, even if the
architecture is unrecognised.  VQE-C has no need for architecture
specific assembly.

The only place I can see where there is assembly is in logging code
for fast timestamping.  A pure C fallback should be provided there
instead but I haven't had time to implement and test this at this
point.