Allow socket sharing between repair RTP and primary RTCP
If the primary stream RTCP feedback target port number has been
set to the RTP port number of the repair stream the same socket
(and thus same src and dest ports) will be used for both.  This
allows NATs to be traversed without requiring STUN or other
complex NAT traversal mechanisms.

This means that any RTCP replies from the server will be
multiplexed with the repair RTP stream.  Currently VQE-C will
ignore (drop) any of these RTCP packets but demuxing could be
added in the future.  This is consistent with RFC 6284 - "Port
Mapping between Unicast and Multicast RTP Sessions".