Prevent VQEC from running out of memory
It seems there's a bug in VQEC that manifests itself with a fact
that it's silently running out of memory (meaning it's not surfaced
to the clinet). This happens despite the fact that the number of active
tuners is actually lower then the maximum specified in the config
file.

This commit fixes that problem by using malloc/free instead of
the custom allocators normally used.