Prevent VQEC from running out of memory
It seems there's a  bug in VQEC  that manifests itself with a fact
that it's running out of memory  surfacing VQEC_ERR_SYSCALL error.
This happens  despite the fact that the number of active tuners is
actually lower then the maximum specified in the config file.

This commit fixes that problem by using malloc/free instead of the
custom allocators normally used. VQEC  allocates all of its memory
upfront, trying  to  prevent   memory  fragmentation, spikes,  and
performance  related   issues. Sadly this  mechanisms fails in our
case when it really should not.