Allocate a random receive port when in sig_mode=std as well as sig_mode=nat
The SDP file is used to communicate what ports the server is using to
the client such that the client can use these ports as destination port
for RTCP messages and local port for connecting to multicast groups.
The SDP file cannot specify which ports the client should use as it
can't know which ports are available and the ports will be changed when
traversing NAT anyway.

For some reason, when sig_mode="std" is set and the primary stream is
a multicast stream the server ports are mirrored in the client.  This is
dangerous as the specified ports may already be in use by some other
process in the system and failure is guaranteed if multiple channels
(e.g. multiple SDP files) specify that the same server port should be
used and more than one of these channels is tuned at any one time.

This commit removes the conditional such that sig_mode="std" behaves the
same as sig_mode="nat" and the client ports are allocated as needed by
the kernel avoiding these issues.  This should be a obviously correct
change but it involves removing two conditionals which seem to have been
very deliberately included.  Unfortunately the comment that used to be
there only describes what the conditionals do rather than why they exist
at all.

This should have no effect for clients configured with sig_mode="nat".